requirements.txt
Flask
psycopg2
redis
kafka-python
confluent-kafka
requests
